+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * UI:
     * New handle for translating layers, groups etc without having to drag on the object
     * Menu items for images in the asset view
+    * Saving the document swatch will ask for a name
 * I/O:
     * Made error messages from the lottie importer easier to understand
 * Bug Fixes: